+++ src/@types/next-auth.d.ts
@@ -4,8 +4,6 @@ import { User } from 'next-auth';
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { JWT } from 'next-auth/jwt';
 
-import { TExtendedUser } from '@/shared/types/TUser';
-
 export type ExtendedUser = User & {
   role: UserRole;
 };