+++ src/components/layout/GenericLayout.tsx
@@ -3,11 +3,12 @@ import { setRequestLocale } from 'next-intl/server';
 
 import { env } from '@/env';
 import { TPropsWithChildren } from '@/shared/types/generic';
-import { TExtendedUser } from '@/shared/types/TUser';
+// import { TExtendedUser } from '@/features/users/types/TUser';
 import { getCurrentUser } from '@/lib/session';
 import { NavBar } from '@/components/layout/NavBar';
 import { NavMobile } from '@/components/layout/NavMobile';
 import { SiteFooter } from '@/components/layout/SiteFooter';
+import { TExtendedUser } from '@/features/users/types/TUser';
 import { TLocale } from '@/i18n/types';
 
 function GenericLayoutContent(props: TPropsWithChildren & { user?: TExtendedUser }) {